# api/internal/server/docs/swagger.yaml
basePath: /
definitions:
  auth.Principal:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture_url:
        type: string
      type:
        $ref: '#/definitions/auth.PrincipalType'
    type: object
  auth.PrincipalType:
    enum:
    - user
    - service_account
    - anyone
    - anonymous
    type: string
    x-enum-varnames:
    - PrincipalUser
    - PrincipalServiceAccount
    - PrincipalAnyone
    - PrincipalAnonymous
  repo.Document:
    properties:
      current_version_id:
        type: string
      id:
        type: string
      owner:
        $ref: '#/definitions/auth.Principal'
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: This API powers a review tool where people can upload one self-contained
    HTML file, view it, and leave comments anchored to specific text ranges inside
    that HTML.
  title: aidocs API
  version: v1
paths:
...
  /v1/auth/cli/credentials:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: List CLI credentials
      tags:
      - auth
  /v1/auth/cli/credentials/{id}:
    delete:
      parameters:
      - description: Credential ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Revoke CLI credential
      tags:
      - auth
  /v1/auth/cli/exchange:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Exchange CLI login code
      tags:
      - auth
  /v1/auth/google/callback:
    get:
      parameters:
      - description: OAuth code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
      summary: Google OAuth callback
      tags:
      - auth
  /v1/auth/google/start:
    get:
      parameters:
      - description: web or cli
        in: query
        name: mode
        type: string
      - description: web redirect
        in: query
        name: redirect
        type: string
      - description: CLI loopback redirect
        in: query
        name: cli_redirect
        type: string
      - description: PKCE challenge
        in: query
        name: code_challenge
        type: string
      - description: OAuth state
        in: query
        name: state
        type: string
      responses:
        "302":
          description: Found
      summary: Start Google OAuth
      tags:
      - auth
...
  /v1/documents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: List documents
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Document title
        in: formData
        name: title
        required: true
        type: string
      - description: Single-file HTML
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Create document
      tags:
      - documents
  /v1/documents/{id}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Delete document
      tags:
      - documents
    get:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Document'
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Get document
      tags:
      - documents
    patch:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Document'
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Update document
      tags:
      - documents
...
  /v1/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      - cookieAuth: []
      summary: Current principal
      tags:
      - auth
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
  cookieAuth:
    in: cookie
    name: aidocs_session
    type: apiKey
swagger: "2.0"